pyproject.toml
[tool.poetry]
name = "hackernewsdiscordfeed"
version = "0.1.0"
description = "A Discord bot for posting articles from the HackerNews web feed"
authors = ["mauvehed <nate@mvh.dev>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
requests = "^2.32.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"
mypy = "^1.10.0"
coverage = "^7.5.2"
types-requests = "^2.32.0.20240523"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["hackernewsdiscordfeed", "tests"]